fix: add time spefication before a reload is done user specified reload before int changed type to int from int64 and only cast in a place where it needs to be converted added unit test

+++ internal/auth/awssts.go
@@ -33,7 +33,7 @@ func LoginStsSaml(samlResponse string, role *util.AWSRole) (*util.AWSCredentials
 		PrincipalArn:    aws.String(role.PrincipalARN), // Required
 		RoleArn:         aws.String(role.RoleARN),      // Required
 		SAMLAssertion:   aws.String(samlResponse),      // Required
-		DurationSeconds: aws.Int64(role.Duration),
+		DurationSeconds: aws.Int64(int64(role.Duration)),
 	}
 
 	resp, err := svc.AssumeRoleWithSAML(params)
